Move rewrite to SQL folder
